Rearrange functions into separate modules
