+++ zaza/openstack/utilities/generic.py
@@ -723,3 +723,22 @@ def get_leaders_and_non_leaders(application_name):
         else:
             non_leaders.append(unit)
     return leader, non_leaders
+
+
+def add_loop_device(unit, size=10):
+    """Add a loopback device to a Juju unit.
+
+    :param unit: The unit name on which to create the device.
+    :type unit: str
+
+    :param size: The size in GB of the device.
+    :type size: int
+
+    :returns: The device name.
+    """
+    loop_name = '/home/ubuntu/loop.img'
+    truncate = 'truncate --size {}GB {}'.format(size, loop_name)
+    losetup = 'losetup --find {}'.format(loop_name)
+    lofind = 'losetup -a | grep {} | cut -f1 -d ":"'.format(loop_name)
+    cmd = "sudo sh -c '{} && {} && {}'".format(truncate, losetup, lofind)
+    return model.run_on_unit(unit, cmd)
